reads and prints info from MOM gradebook csv file
This program will read in a csv file of my actual gradebook downloaded from myOpenMath.com. It will then loop through each student and print the name and grade of anyone with a grade above 90%. This will need modification so that instead of conditioning on the weighted average, it conditions on the number of absences. Also instead of printing the student name and weighted average, we want it to do several other things